refactor: シーン構成をAppから分離
